Flush out location a bit more!
This commit adds two new API calls:
1. ast_sip_location_retrieve - This retrieves a named location but takes into account location expiration if applicable
2. ast_sip_location_update - This updates the URI and expiration of a location if permitted

It also turns location into a sort of hybrid object. It can exist in both a static configuration file but parts of it
can also be stored in a persistence mechanism, like a database.

Next up I'll be writing a handler for the expiration time and also poking at dialing locations some more.