Get rid of the ast_ wrappers for cloning and freeing pjsip_rx_data.
It was a decent enough idea, but actually making it bug free is going
to be difficult. Also, giving the distributor knowledge of what is stored
in the rdata's mod_data array breaks what little encapsulation already
exists, so it just feels very dirty. It may be possible to make things
more intricate, but the benefit is not that great.

The result now is that the endpoint needs to be retrieved from the
rdata before pushing a task to the serializer. This is not really
all that inconvenient.