Edit documentation explaining how incoming messages are distributed to servants.
This explains that incoming requests and responses are automatically distributed
to servant threads.

While writing the documentation I realized that there was no easy way to discern
whether you are currently in a servant thread or not. So I've modified SIP thread
startup in order to set a thread-local identifier. Then I added a function called
ast_sip_thread_is_servant() that can be used to find if the current thread is
a SIP servant thread. It is not used anywhere yet since the only PJSIP callbacks
that may be called have to be unconditionally pushed to serializers anyway.
However, it could be useful in certain cases.