Move serialization to the distributor.
This adds support for specifying a serializer on a per-dialog basis. When a request
or response comes in the dialog is found and the serializer used. This also means that
all modules involved in handling a message are invoked within the serializer, and not
partially inside and outside. As a result in most cases you no longer have to push your
own tasks.