Add options to determine how endpoints are identified.
Prior to this commit, the method by which endpoints were identified was
dependent on what endpoint identification modules were loaded. While this
works, it's not likely to be useful for most deployments. Instead, having
control within the configuration of how identification should occur per
endpoint is more useful.

There is now an "identify_by" option that accepts a comma-delimited list of
methods by which the endpoint can be identified. By default, this is set to
"username,location" which allows for the endpoint to be identified by both
by username and by its IP address. The option can be changed to only allow
one method or the other.

Using this allows for easy expansion in case new methods of endpoint identification
are devised.