Set up outbound authentication for SIP sessions.
This creates the authentication options necessary for setting
up outbound authentication. Now in addition to "auth=" for endpoints,
there is also "outbound_auth=" that can be set. These are used in case
a request to that endpoint is challenged.

In res_sip_session.c, we now set up authentication credentials on
the dialog. This *should* have the effect of having challenges be
automatically handled by the dialog.

A more generic method will need to be devised for situations where
no dialog is in use.