Progress on direct media work.
For now, I'm operating with one simple setting: Allow or don't allow
direct media.

In my tests, there are some issues:

1) Transmission of the reinvites to the endpoints is inconsistent.
Sometimes no reinvites are sent. Sometimes a reinvite is sent to one
party. Sometimes reinvites are sent to both parties.

2) In the case where any reinvites are sent, then when the call ends,
Asterisk crashes. It appears to be due to the fact that a reinvite is
sent to a party immediately followed by a BYE. This makes it so that
when the 200 OK for the reinvite arrives, we access data that has been
released already.