More progress towards direct media calls.
I've taken care of the crash seen during hangup by ensuring
we do not try to apply a negotiated SDP if there is no channel
to apply it to.

There is still a race condition that sometimes causes reinvites
not to get sent on the outbound leg. It happens about one in five
times during a call and likely has something to do with the timing
of when the RTP layer is calling the update_peer glue function.

I think the issue is that the command to send a reinvite is happening
between when the INVITE transaction terminates and when the inv_session
layer actually clears the invite_tsx from itself. This means that we
delay sending the reinvite due to a pending transaction, but we never
get told that the transaction has cleared since there are no more
transaction state changes. Instead, we may have to rely on a signal
from the inv_session layer to indicate that it has cleared its
invite_tsx.