Fix the reinvite race condition I was experiencing.
As my last commit message stated, I was seeing times where
the reinvite that should be sent to the outbound leg was never
being sent.

The reason was because the inv_session telling when a transaction
was terminated did not correspond to when the inv_session cleared its
invite_tsx field. However, by changing when I send delayed requests
to the transaction layer's transaction state change callback, this
corresponds well with the inv_session's clearing of the invite_tsx.

I currently still have the delayed request being queued into the
serializer, but I believe I can get rid of that since I should already
be serialized and there is nothing that I need to have between getting
the transaction state change notice and sending the delayed request.