Don't send a premature BYE when a transaction state is in "Completed" state.
For an INVITE transaction in the UAC role, the "Completed" state occurs when
a failure response (>=300) is received. However, in the UAR role, the "Completed"
state occurs between sending a final response (whether it is 200 or not) and
reception of the ACK.

For the purposes of reinvites, we only want to terminate the session when we
receive an error response for an INVITE we sent out.