Implement a callback system for session responses.
Now when an entity calls for a request to be sent, they can have a callback
called when the final response for their request arrives.

In addition, the session refresh method has a callback that can be called
when the request is created. This allows for a caller to add custom headers
if there is a need to do so before sending the request.