Put code in place for direct media glare mitigation.
A config option, aptly called direct_media_glare_mitigation,
directs us how we can attempt to reduce reinvite glare. We
can either do nothing, not send an initial refresh on outgoing
calls, or not send an initial refresh on incoming calls.

In order to only allow for this to occur once per call, a datastore
is set up on the session during session creation. The datastore is
then removed when a reinvite is prevented from being sent out.