Fix deadlock when smart bridge destroys the softmix bridge structure.
The technology bridge structure for softmix has a mixing thread that can
hold the bridging lock.  It cannot be stopped when that lock is held by
the thread trying to stop the mixing thread.

Since the smart bridge operation cannot release the bridge lock, the
solution is to defer/pass destroying the technology bridge structure to
the bridge manager thread.  It can then destroy the structure with the
bridge not locked.