Separate request creation from sending.
This allows for custom headers to be added easily to SIP requests. It
also allows flexibility in creation and sending of requests. A request
could be created via a PJSIP API call and then sent using the Asterisk
one, or vice versa.

Tested and it works. All that's necessary now is some cleanup and it'll
be review time!