Don't look at this.  It's horrible.
It totally works though.  Sort of.  Except there's no SRTP yet.  And the
threading is completely broken, and will be until I bother to look at how
that's all supposed to work.

It can register and calls can be started.  Sometimes it'll even hangup without
crashing, too.  Not frequently though.

Also, no IPv6 support is possible, because of a flaw in how custom transport
types get registered in pjproject.