The configure script in this branch was broken.
A macro from the auto-configure script did not expand properly
and caused the configure script to fail.

I've regenerated the configure script here. This will almost
certainly cause a conflict the next time automerge occurs. When
that happens, we'll respond by just using the version of the
configure script that was being merged in.