Initial Caller ID support.
This implements a simple supplement that will set caller ID on the session
on an incoming INVITE and set ID details on an outgoing session based on
channel data.

It's a pretty simple setup and it works quite well if the Dial 'o' option
is used since I have no connected line support in yet.

To make this work accurately, I've introduced a crude priority system for
session supplements so that we can ensure that they register in a proper
order even if modules load in unexpected intervals. This way I can guarantee
that on incoming INVITEs, the caller ID module is visited before chan_gulp.
This will need to be expanded to use some constants so that the priorities
make a bit more sense than they do at the moment.