Reorg RESTful API to get swagger-codgen and swagger-ui both happy.
Swagger-codegen assumes that the first element in the path of a resource's
path is the class name for the resource. All of ours started with /api,
which meant that everything landed in a single ApiApi class.

So instead of putting the docs in /stasis and the API in /stasis/api,
we now have the API in /stasis and the docs i /stasis/api-docs. The
sources were also rearranged so that codegen can be run from the source
without running Asterisk.