confbridge: Fix announcer channel bridge depart getting stuck.
When the announcer channel needs to depart from the bridge, the thread
needs to be poked to get it out of an indefinite wait for media.
Unfortunately, this fix does not guarantee that ast_bridge_depart() will
not get stuck waiting for the bridge channel thread to die.  Thres is a
small window of opportunity in bridge_channel_wait() where the thread is
about to enter the ast_waitfor_nandfds() when the thread poke will be
ignored.
........

Merged revisions 385255 from file:///srv/subversion/repos/asterisk/team/group/bridge_construction