confbridge: Fix announcer channel bridge depart getting stuck.
When the announcer channel needs to depart from the bridge, the thread
needs to be poked to get it out of an indefinite wait for media.  More
generally, the bridge channel thread needs to be poked whenever it is in
the normal bridge_channel_wait()'s ast_waitfor_nandfds() call and the
state changes.

* Fixes the small window of opportunity in bridge_channel_wait() when
entering ast_waitfor_nandfds().
........

Merged revisions 385272 from file:///srv/subversion/repos/asterisk/team/group/bridge_construction