Correct refcounting for solicited MWI.
The references work like this:

Stasis holds a reference to the MWI subscription, which
holds a reference to the pubsub subscription. When the SIP
subscription terminates, we unsubscribe from stasis. When
stasis sends its final message to us, we can then lose stasis's
reference to the MWI subscription, which causes destruction
down the chain.

With this improvement, MWI is functional. Remaining tasks
are just to do code cleanup and documentation.