This adds support for outbound authentication for outbound registrations.
In order to facilitate the change, the outbound authentication API had to
be modified not to be dependent on endpoints but rather to be given the
names of auth IDs in sorcery to use. This way, outbound registration
can be configured to point to an auth section in the configuration the
same way that endpoints do.

Review: https://reviewboard.asterisk.org/r/2431
........

Merged revisions 385967 from file:///srv/subversion/repos/asterisk/team/group/pimp_my_sip