Create the basic bridge subclass of struct ast_bridge.
This bridge class is the normal bridge for most calls.  It adds management
of the DTMF feature hooks for each channel while they are in this bridge.
When the basic two party bridge expands into a multi-party bridge, the
various parties can drop out of the bridge without destroying it.

The new bridging API does not allow casual knowledge of what channels are
bridged with a particular channel because there can be more than one.
With this restriction, the support of the [applicationmap] ActivatedBy
feature option no longer makes sense and cannot be supported.  The
ActivatedBy option is now just ignored.  The feature can only be activated
by the channel it is attached to and not by its peer.

(closes issue ASTERISK-21332)
Reported by: Matt Jordan

Review: https://reviewboard.asterisk.org/r/2446/