Get local channel optimization working again using the new bridging model.
The primary method of optimization is to merge bridges together.  The
optimizing local channels continuously check for empty queues in the
local;1 and local;2 channels.  Once it is found that the queues are empty
and the bridges allow merging, either BridgeA or BridgeB is merged into
the other bridge.

A future additional way to optimize out local channels in more
circumstances is to swap the Local;2 channel in BridgeB with the only peer
channel (A) in BridgeA.  This optimization strategy would be needed if
BridgeB prohibits bridge merges.

A -- BridgeA -- Local;1-Local;2 -- BridgeB -- B

(closes issue ASTERISK-21058)
Reported by: Matt Jordan

Review: https://reviewboard.asterisk.org/r/2461/