Create a new branch for bridge transfer work.
The idea behind this is to create a few API calls that
channel driver and other "external" entities can call in
order to perform call redirection, such as blind and attended
transfers.

Since channel masquerades are still very much necessary for
certain transfer scenarios, it means that they will be present
in this transfer code. However, instead of having multiple
calls to ast_channel_masquerade/ast_do_masquerade in different
transfer functions, the calls to these functions will be isolated
to a single method. The idea is that (with the possible exception
of local channel optimization) it may be possible to have a single
place in all of Asterisk that masquerades are performed.