Use the ast_bridge_peers() function to get the transferee.
This probably should be done at a different time since the current
construct counts the channels in the bridge, then unlocks the bridge
and then gets the channels in the bridge. There exists the opportunity
for an unwanted third party to join the bridge during that gap.

Currently, an attempt to use the BlindTransfer manager action results
in a crash.