Fix crash and queue bridge action on expected bridge channel.
The crash does not happen, but the blind transfer still does not work.
The issue is that by changing the bridge channel state on the transferee
channel, the result is that the channel itself is hung up on, and the
other channel in the bridge remains, alone.

The expecations are that

1) The transferee channel is removed from the bridge and executes dialplan
where the explicit goto says to do so.
2) The transferer channel, being the only one remaining in a two-party bridge
will be hung up on.

If I change the transferee state to be END instead of HANGUP, the bridge
does dissolve, but the transferee channel still gets hung up instead of
continuing on.