Leave removal of the transferer from the bridge to the bridge itself.
The next item on the agenda is to add the special case of performing a
blind transfer to a parking extension. In order to facilitate this, we
can't make the assumption that the transferer is supposed to be hung
up upon return of ast_bridge_blind_transfer(). It may be that the transferer
is supposed to remain alive until a parking announcement is made.

Tested this, and blind transfers still work properly in both two-party
bridges and confbridges.