Make use of the framehook passed into ast_bridge_blind_transfer()
Whether transferring the entire bridge or a single channel, the framehook
is applied to the outbound channel that is entering the dialplan. This
way, if the caller of ast_bridge_blind_transer() cares about the progress
of that call, they can track it via the framehook.

I also fixed an incorrect piece of documentation in the framehook API.
"positive" and "non-negative" are not the same thing.