Refactor manager bridge action to be more bridge-y.
This makes use of the channel's current bridge disposition and
simply moves the channel if possible. Otherwise, the channel is
yanked and placed into the new bridge by force.

Currently, there are a couple of issues with this:

1) Refcounts are all over the place. I can probably get away with
a couple of well-placed RAII_VAR macros.
2) I'm currently not playing a courtesy tone to the channel if it
is bridged. This probably can be done by queuing a bridge action either
before or after moving the channel into the new bridge.