Fix bridge logic in SIP's handle_invite_replaces()
In my tests, bridged scenarios were working, but as Josh
pointed out, it turns out that the reason for this is because
the bridged scenario could not actually evaluate true. This
meant that in all cases, I ended up performing an ast_channel_move()
instead of an ast_bridge_impart().