Fix a few things...
First, add SWAP_INHIBIT_TO since local channels were able to take the place of
a local channel that was parked which causes interval hooks to fall off and the
local channel to be able to occupy the parking space indefinitely. There might
be a better fix for that down the road.

Second, destroy an iterator that was accidentally left alive when doing one touch
parking and transfers to parking lots. This would leak channel references on both
the parker and parkee side when doing two party parking which prevented channel
destruction.

Third, fix a problem leading to a segfault when tearing down parker announcement
subscriptions. Also free the datastore which manages these subscriptions when the
datastore destroy function is invoked.