Create public function for getting a channel's applicationmap.
This function examines the DYNAMIC_FEATURES variable on a channel in
order to determine what features are available for the channel and
returns an ao2_container with each of the enabled features.

There is currently the possibility for duplicate entries to be
placed in the returned container, but that will be addressed later.