Package common attended transfer data into a "publication" structure.
This gives us two distinct advantages:

1) It allows for a XXX comment to be removed. Now we are guaranteed to indicate
the correct parties when attended transferring a bridge.

2) It allows for an upcoming improvement to be made easily. We want to be consistent
about when we publish the transfer. When performing an attended transfer of a single
party to an application, we currently do so by queuing a bridge action. This action
runs in a separate thread. If we pass the publication data as part of the action data,
then we can hold off on publishing the stasis message until the transfer has actually
completed. Note: this also applies for blind transfers.

The one problem with this is that I've simply tacked on the publication to function calls.
The publication likely should just replace parameters since the parameters are somewhat redundant.
Also the publication parameters have not been added to the doxygen.